resolve PR comments part1
