+++ public/app/features/alerting/unified/components/receivers/EditReceiverView.tsx
@@ -6,7 +6,7 @@ import { GRAFANA_RULES_SOURCE_NAME } from '../../utils/datasource';
 import { CloudReceiverForm } from './form/CloudReceiverForm';
 import { GrafanaReceiverForm } from './form/GrafanaReceiverForm';
 
-export interface EditReceiverViewProps {
+interface EditReceiverViewProps {
   alertmanagerName: string;
   contactPoint: GrafanaManagedContactPoint | Receiver;
 }
