+++ public/app/features/alerting/unified/components/receivers/form/GrafanaReceiverForm.tsx
@@ -100,28 +100,23 @@ export const GrafanaReceiverForm = ({ contactPoint, readOnly = false, editMode }
   const onSubmit = async (values: ReceiverFormValues<GrafanaChannelValues>) => {
     const newReceiver = formValuesToGrafanaReceiver(values, id2original, defaultChannelValues);
 
-    try {
-      if (editMode) {
-        if (contactPoint && contactPoint.id) {
-          await updateContactPoint.execute({
-            contactPoint: newReceiver,
-            id: contactPoint.id,
-            resourceVersion: contactPoint?.metadata?.resourceVersion,
-          });
-        } else if (contactPoint) {
-          await updateContactPoint.execute({
-            contactPoint: newReceiver,
-            originalName: contactPoint.name,
-          });
-        }
-      } else {
-        await createContactPoint.execute({ contactPoint: newReceiver });
+    if (editMode) {
+      if (contactPoint && contactPoint.id) {
+        await updateContactPoint.execute({
+          contactPoint: newReceiver,
+          id: contactPoint.id,
+          resourceVersion: contactPoint?.metadata?.resourceVersion,
+        });
+      } else if (contactPoint) {
+        await updateContactPoint.execute({
+          contactPoint: newReceiver,
+          originalName: contactPoint.name,
+        });
       }
-      locationService.push('/alerting/notifications');
-    } catch (error) {
-      // Propagate so ReceiverForm can show notifyApp.error with the backend message
-      throw error;
+    } else {
+      await createContactPoint.execute({ contactPoint: newReceiver });
     }
+    locationService.push('/alerting/notifications');
   };
 
   const onTestChannel = (values: GrafanaChannelValues) => {
