+++ public/app/features/alerting/unified/triage/instance-details/ContactPointInstanceDrawerToolbar.tsx
@@ -1,12 +1,13 @@
 import { Fragment, type JSX, useState } from 'react';
 
-import { Trans, t } from '@grafana/i18n';
-import { config } from '@grafana/runtime';
-import { Dropdown, LinkButton, Menu, Stack, Tooltip } from '@grafana/ui';
-import ConditionalWrap from 'app/features/alerting/unified/components/ConditionalWrap';
+import { t } from '@grafana/i18n';
+import { Dropdown, LinkButton, Menu, Stack } from '@grafana/ui';
 import MoreButton from 'app/features/alerting/unified/components/MoreButton';
 import { ProvisioningBadge } from 'app/features/alerting/unified/components/Provisioning';
 import { Spacer } from 'app/features/alerting/unified/components/Spacer';
+import { ContactPointDeleteMenuItem } from 'app/features/alerting/unified/components/contact-points/components/ContactPointDeleteMenuItem';
+import { ContactPointHistoryLinkButton } from 'app/features/alerting/unified/components/contact-points/components/ContactPointHistoryLinkButton';
+import { ContactPointManagePermissionsMenuItem } from 'app/features/alerting/unified/components/contact-points/components/ContactPointManagePermissionsMenuItem';
 import { UnusedContactPointBadge } from 'app/features/alerting/unified/components/contact-points/components/UnusedBadge';
 import { useExportContactPoint } from 'app/features/alerting/unified/components/contact-points/useExportContactPoint';
 import {
@@ -66,11 +67,7 @@ export function ContactPointInstanceDrawerToolbar({ contactPoint, onDelete }: Co
   if (showManagePermissions) {
     menuActions.push(
       <Fragment key="manage-permissions">
-        <Menu.Item
-          icon="unlock"
-          label={t('alerting.contact-point-header.label-manage-permissions', 'Manage permissions')}
-          onClick={() => setShowPermissionsDrawer(true)}
-        />
+        <ContactPointManagePermissionsMenuItem onOpen={() => setShowPermissionsDrawer(true)} />
       </Fragment>
     );
   }
@@ -92,61 +89,18 @@ export function ContactPointInstanceDrawerToolbar({ contactPoint, onDelete }: Co
   }
 
   if (deleteSupported) {
-    const cannotDeleteNoPermissions = t(
-      'alerting.contact-points.delete-reasons.no-permissions',
-      'You do not have the required permission to delete this contact point'
-    );
-    const cannotDeleteProvisioned = t(
-      'alerting.contact-points.delete-reasons.provisioned',
-      'Contact point is provisioned and cannot be deleted via the UI'
-    );
-    const cannotDeletePolicies = t(
-      'alerting.contact-points.delete-reasons.policies',
-      'Contact point is referenced by one or more notification policies'
-    );
-    const cannotDeleteRules = t(
-      'alerting.contact-points.delete-reasons.rules',
-      'Contact point is referenced by one or more alert rules'
-    );
-
-    const reasonsDeleteIsDisabled = [
-      !hasAbilityToDelete ? cannotDeleteNoPermissions : '',
-      isProvisioned ? cannotDeleteProvisioned : '',
-      numberOfPoliciesPreventingDeletion > 0 ? cannotDeletePolicies : '',
-      numberOfRules ? cannotDeleteRules : '',
-    ].filter(Boolean);
-
-    const deleteTooltipContent = (
-      <>
-        <Trans i18nKey="alerting.contact-points.delete-reasons.heading">
-          Contact point cannot be deleted for the following reasons:
-        </Trans>
-        <br />
-        {reasonsDeleteIsDisabled.map((reason) => (
-          <li key={reason}>{reason}</li>
-        ))}
-      </>
-    );
-
     menuActions.push(
-      <ConditionalWrap
-        key="delete-contact-point"
-        shouldWrap={!canBeDeleted}
-        wrap={(children) => (
-          <Tooltip content={deleteTooltipContent} placement="top">
-            <span>{children}</span>
-          </Tooltip>
-        )}
-      >
-        <Menu.Item
-          label={t('alerting.contact-point-header.label-delete', 'Delete')}
-          ariaLabel={t('alerting.contact-point-header.ariaLabel-delete', 'Delete')}
-          icon="trash-alt"
-          destructive
-          disabled={!canBeDeleted}
-          onClick={() => onDelete(contactPoint)}
+      <Fragment key="delete-contact-point">
+        <ContactPointDeleteMenuItem
+          contactPoint={contactPoint}
+          onDelete={onDelete}
+          canBeDeleted={canBeDeleted}
+          hasAbilityToDelete={hasAbilityToDelete}
+          isProvisioned={isProvisioned}
+          numberOfPoliciesPreventingDeletion={numberOfPoliciesPreventingDeletion}
+          numberOfRules={numberOfRules}
         />
-      </ConditionalWrap>
+      </Fragment>
     );
   }
 
@@ -178,28 +132,8 @@ export function ContactPointInstanceDrawerToolbar({ contactPoint, onDelete }: Co
     </LinkButton>
   );
 
-  const historyButton = config.featureToggles.alertingNotificationHistoryGlobal && (
-    <LinkButton
-      variant="secondary"
-      size="sm"
-      icon="history"
-      type="button"
-      disabled={integrations.length === 0}
-      tooltip={t(
-        'alerting.contact-point-header.tooltip-history',
-        'View the history of notification attempts made to this contact point'
-      )}
-      tooltipPlacement="top"
-      data-testid="history-action"
-      href={createRelativeUrl('/alerting/history', {
-        tab: 'notifications',
-        'var-RECEIVER_FILTER': name,
-      })}
-      target="_blank"
-      rel="noopener noreferrer"
-    >
-      {t('alerting.contact-point-header.button-history', 'History')}
-    </LinkButton>
+  const historyButton = (
+    <ContactPointHistoryLinkButton contactPointName={name} integrationsCount={integrations.length} />
   );
 
   const moreMenu =
